api-gateway/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv>=1.1.1
pydantic>=2.11.9
